crossdock - Move Call() to crossdock-go
