+++ crossdock/crossdock_test.go
@@ -21,27 +21,15 @@
 package main
 
 import (
-	"encoding/json"
-	"log"
 	"net/url"
 	"testing"
-	"time"
 
 	"github.com/yarpc/yarpc-go/crossdock-go/crossdock"
 	"github.com/yarpc/yarpc-go/crossdock/client"
 	"github.com/yarpc/yarpc-go/crossdock/server"
-
-	"github.com/stretchr/testify/require"
-	"golang.org/x/net/context"
-	"golang.org/x/net/context/ctxhttp"
 )
 
 const clientURL = "http://127.0.0.1:8080"
-
-type result struct {
-	Status string `json:"status"`
-	Output string `json:"output"`
-}
 
 func TestCrossdock(t *testing.T) {
 	server.Start()
@@ -100,13 +88,12 @@ func TestCrossdock(t *testing.T) {
 	for _, bb := range behaviors {
 
 		args := url.Values{}
-		args.Set("behavior", bb.name)
 		for k, v := range defaultParams {
 			args.Set(k, v)
 		}
 
 		if len(bb.axes) == 0 {
-			call(t, args)
+			crossdock.Call(t, clientURL, bb.name, args)
 			continue
 		}
 
@@ -122,32 +109,7 @@ func TestCrossdock(t *testing.T) {
 				entryArgs.Set(k, v)
 			}
 
-			call(t, entryArgs)
-		}
-	}
-}
-
-func call(t *testing.T, args url.Values) {
-	u, err := url.Parse("http://127.0.0.1:8080")
-	require.NoError(t, err, "failed to parse URL")
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	u.RawQuery = args.Encode()
-	log.Println("GET", u.String())
-	res, err := ctxhttp.Get(ctx, nil, u.String())
-
-	require.NoError(t, err, "request %v failed", args)
-	defer res.Body.Close()
-
-	var results []result
-	require.NoError(t, json.NewDecoder(res.Body).Decode(&results),
-		"failed to decode response for %v", args)
-
-	for _, result := range results {
-		if result.Status != "passed" && result.Status != "skipped" {
-			t.Errorf("request %v failed: %s", args, result.Output)
+			crossdock.Call(t, clientURL, bb.name, entryArgs)
 		}
 	}
 }
